pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
DateTime = "^4.4"
discord = "^1.7.3"
Flask = "^2.1.2"
ipc = "^1.2.1"
requests = "^2.27.1"
youtube_dl = "^2021.12.17"
timestamp = "^0.0.1"
"dislash.py" = "^1.4.9"
mcstatus = "^9.0.4"
PyNaCl = "^1.5.0"
easy-pil = "^0.1.6"
pytube3 = "^9.6.4"
ffmpeg = "^1.4"
webserver = "^0.10.3"
utils = "^1.0.1"
colorama = "^0.4.5"
pytz = "^2022.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"